fix type checker

+++ simplyblock_web/api/internal/storage_node/kubernetes.py
@@ -492,6 +492,9 @@ def spdk_process_kill(query: utils.RPCPortParams):
     k8s_core_v1 = core_utils.get_k8s_core_client()
     try:
         namespace = node_utils_k8s.get_namespace()
+        if not query.cluster_id:
+            return utils.get_response(False, f"param required: cluster_id")
+
         first_six_cluster_id = core_utils.first_six_chars(query.cluster_id)
         pod_name = f"snode-spdk-pod-{query.rpc_port}-{first_six_cluster_id}"
         resp = k8s_core_v1.delete_namespaced_pod(pod_name, namespace)
